@brokensound77 2023-03-07
> LGTM. From my understanding, there is two enhancements here. The first, we use the `@lru_cache` decorator for the rule loader, which loads all rules. This way all rules are not loaded multiple times as it is already cached. Second, the `setUp` method takes the value of `RULE_LOADER_FAIL` which is set by `setUpClass` and if `True` stops the rule loader from doing additional runs.
> 
> @brokensound77 please correct my understanding if it is incorrect.

yep, exactly. Any class that inherits the `BaseRuleTest`, which is all of the rule test cases, will be auto skipped on failure, since there are no rules to test. Before, it would try to reload the rules every time